Move exceptions to handle cyclic import
